Refactor acquisition_metadata.
The acquisition_metadata is now independent of the metaseries metadata style.
