+++ tests/test_Zarr.py
@@ -5,8 +5,10 @@ import unittest
 from os.path import exists, join
 from pathlib import Path
 
-from faim_hcs.ImageUtils2D import get_well_image_CYX
+import numpy as np
+
 from faim_hcs.io.MolecularDevicesImageXpress import parse_single_plane_multi_fields
+from faim_hcs.MetaSeriesUtils import get_well_image_CYX
 from faim_hcs.Zarr import build_zarr_scaffold, write_cyx_image_to_well
 
 ROOT_DIR = Path(__file__).parent
@@ -99,7 +101,7 @@ class TestZarr(unittest.TestCase):
             )
         )
         assert "histograms" in c05.keys()
-        assert "metaseries_metadata" in c05.keys()
+        assert "acquisition_metadata" in c05.keys()
         assert c05["multiscales"][0]["datasets"][0]["coordinateTransformations"][0][
             "scale"
         ] == [1.0, 0.3417, 0.3417]
@@ -116,10 +118,49 @@ class TestZarr(unittest.TestCase):
             )
         )
         assert "histograms" in c06.keys()
-        assert "metaseries_metadata" in c06.keys()
+        assert "acquisition_metadata" in c06.keys()
         assert c06["multiscales"][0]["datasets"][0]["coordinateTransformations"][0][
             "scale"
         ] == [1.0, 0.3417, 0.3417]
+
+    def test_get_well_image_CYX(self):
+        cyx, hists, ch_meta, metadata = get_well_image_CYX(
+            well_files=self.files[self.files["well"] == "C05"]
+        )
+
+        assert cyx.shape == (2, 2048, 4096)
+        assert cyx.dtype == np.uint16
+        assert len(hists) == 2
+        assert hists[0].min() == 107
+        assert hists[1].min() == 126
+        assert ch_meta[0] == {
+            "channel-name": "DAPI",
+            "display-color": "0034ff",
+            "exposure-time": 20.0,
+            "exposure-time-unit": "ms",
+            "objective": "20X Plan Apo Lambda",
+            "objective-NA": 0.75,
+            "power": 100.0,
+            "shading-correction": False,
+            "wavelength": "violet",
+        }
+        assert ch_meta[1] == {
+            "channel-name": "FITC",
+            "display-color": "73ff00",
+            "exposure-time": 250.0,
+            "exposure-time-unit": "ms",
+            "objective": "20X Plan Apo Lambda",
+            "objective-NA": 0.75,
+            "power": 100.0,
+            "shading-correction": False,
+            "wavelength": "cyan",
+        }
+        assert metadata == {
+            "pixel-type": "uint16",
+            "spatial-calibration-units": "um",
+            "spatial-calibration-x": 0.3417,
+            "spatial-calibration-y": 0.3417,
+        }
 
 
 if __name__ == "__main__":
